bot/requirements.txt
python-telegram-bot==12.8
Flask==1.1.2
